fix: 전체 코드 리뷰 보안·타입·성능 24건 수정
Phase 1 — 보안/안정성:
- 라우터 4개 _get_device_id() 안전 추출 패턴 적용
- 프로덕션 CORS에서 localhost 제거
- Rate limiter X-Forwarded-For 지원
- 세션쿠키 secure 플래그 환경별 분기
- deploy.sh 마이그레이션 실패 처리

Phase 2 — 타입 안전성:
- Pydantic 스키마 from_attributes + datetime 타입
- Drug/Supplement ingredients dict→list 수정
- Reminder days_of_week BigInteger→Integer
- DB 엔진 풀 + Redis 풀 튜닝

Phase 3 — 성능/품질:
- AI 설명 생성 실패 graceful fallback
- Flutter 검색 순차→병렬 Future.wait 최적화
- Go Router 안전 타입 캐스팅
- Web SearchInput 접근성 + API client SSR 호환

테스트: Backend 229 + Web 114 + Flutter 131 = 474개 전부 통과

+++ src/backend/models/reminder.py
@@ -2,7 +2,7 @@
 
 from datetime import time
 
-from sqlalchemy import BigInteger, Boolean, Index, String, Time
+from sqlalchemy import BigInteger, Boolean, Index, Integer, String, Time
 from sqlalchemy.dialects.postgresql import ARRAY
 from sqlalchemy.orm import Mapped, mapped_column
 
@@ -30,7 +30,7 @@ class Reminder(Base, TimestampMixin):
     cabinet_item_id: Mapped[int] = mapped_column(BigInteger, nullable=False)
     item_name: Mapped[str] = mapped_column(String(500), nullable=False)
     reminder_time: Mapped[time] = mapped_column(Time, nullable=False)
-    days_of_week: Mapped[list[int]] = mapped_column(ARRAY(BigInteger), nullable=False)
+    days_of_week: Mapped[list[int]] = mapped_column(ARRAY(Integer), nullable=False)
     is_active: Mapped[bool] = mapped_column(Boolean, default=True, nullable=False)
     memo: Mapped[str | None] = mapped_column(String(500))
